bugfix harmonic driver
