+++ drivers/drivers/harmonic.py
@@ -23,8 +23,8 @@ class Harmonic_driver(Dummy_driver):
             sys.exit(self.error_msg)
 
     def __call__(self, cell, pos):
-        """ Silly harmonic potential, with unit frequency in a.u."""
-        pot = 0.5 * self.k * (pos ** 2).sum() * 0.5
+        """ Silly harmonic potential"""
+        pot = 0.5 * self.k * (pos ** 2).sum() 
         force = -self.k * pos  # makes a zero force with same shape as pos
         vir = cell * 0.0  # makes a zero virial with same shape as cell
         extras = "nada"
