+++ examples/MBPOL/splitting/instanton/32-lbfgs/test_settings.dat
@@ -1,3 +1,4 @@
 #Settings file for automatic test
-driver qtip4pf
+driver_code fortran
+driver_model qtip4pf
 nsteps 10
