+++ examples/fhi-aims/zundel/test_settings.dat
@@ -1,3 +1,3 @@
 #Settings file for automatic test
-driver zundel
+driver_model zundel
 nsteps 10
