+++ examples/lammps/h2o-imf/test_settings.dat
@@ -1,3 +1,3 @@
 #Settings file for automatic test
-driver qtip4pf
+driver_model qtip4pf
 nsteps 10
