+++ examples/lammps/ice-npzt/test_settings.dat
@@ -1,3 +1,3 @@
 #Settings file for automatic test
-driver harm3d
+driver_model harm3d
 flags -o 2
