+++ examples/pes/pswater/nvt/test_settings.dat
@@ -1,3 +1,3 @@
 #Settings file for automatic test
-driver pswater
+driver_model pswater
 nsteps 10
