+++ examples/ph2/nst-cubic/test_settings.dat
@@ -1,3 +1,3 @@
 #Settings file for automatic test
-driver sg
+driver_model sg
 flags -o 15.0
